Replace `size_t`s with `int64_t`'s in the source and header files relating to `RegularPolygon2D`. (#1)

+++ src/regular_polygon2d.cpp
@@ -15,12 +15,12 @@ void RegularPolygon2D::_bind_methods() {
     ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "radius"), "set_radius", "get_radius");
 }
 
-void RegularPolygon2D::set_nsides(size_t nsides_p) {
+void RegularPolygon2D::set_nsides(int64_t nsides_p) {
     nsides = nsides_p;
     update_polygon();
 }
 
-size_t RegularPolygon2D::get_nsides() {
+int64_t RegularPolygon2D::get_nsides() {
     return nsides;
 }
