+++ src/regular_polygon2d.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #include <godot_cpp/classes/polygon2d.hpp>
 
 using namespace godot;
@@ -7,7 +8,7 @@ using namespace godot;
 class RegularPolygon2D : public Polygon2D {
     GDCLASS(RegularPolygon2D, Polygon2D);
 
-    size_t nsides;
+    int64_t nsides;
     double radius;
 
 protected:
@@ -15,8 +16,8 @@ protected:
 
 public:
 
-    void set_nsides(size_t nsides_p);
-    size_t get_nsides();
+    void set_nsides(int64_t nsides_p);
+    int64_t get_nsides();
 
     void set_radius(double radius_p);
     double get_radius();
